refactor: move cli command logic
- remove ifelse in __main__
- rename api folder to cli
- prepare decoupling of cli from getdeck lib
