+++ getdeck/cli/console.py
@@ -1,0 +1,17 @@
+import logging
+import sys
+
+logger = logging.getLogger("deck")
+
+
+def console_setup(debug: bool = False):
+    console = logging.StreamHandler(sys.stdout)
+    formatter = logging.Formatter("[%(levelname)s] %(message)s")
+    console.setFormatter(formatter)
+
+    if debug:
+        logger.setLevel(logging.DEBUG)
+    else:
+        logger.setLevel(logging.INFO)
+
+    logger.addHandler(console)
