+++ getdeck/cli/version.py
@@ -1,0 +1,10 @@
+import logging
+
+
+logger = logging.getLogger("deck")
+
+
+def version_command(args):
+    from getdeck import configuration
+
+    logger.info(f"Deck version: {configuration.__VERSION__}")
